feat(storage-provider): Move pallet module back to the root and tweak imports
