+++ pallets/storage-provider/benchmarks/src/lib.rs
@@ -5,8 +5,11 @@ extern crate alloc;
 #[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
 
-#[cfg(all(feature = "runtime-benchmarks", test))]
+#[cfg(any(feature = "runtime-benchmarks", test))]
 pub(crate) mod mock;
 
-#[cfg(feature = "runtime-benchmarks")]
-pub use benchmarking::pallet::*;
+#[cfg(any(feature = "runtime-benchmarks", test))]
+mod pallet;
+
+#[cfg(any(feature = "runtime-benchmarks", test))]
+pub use pallet::*;
